Part 3
1. In the root of the task, create a vo (value object) package and create a JobPosting class in it.
This class will store job posting data.

2. In Provider, create a List<JobPosting> getJavaJobPostings(String searchString) method. Leave the method empty for now.

3. What data is associated with a job posting?
A job title, city, company name, name of the website where the job posting was found, and a link to the job posting.
In the JobPosting class, create the corresponding String fields: title, city, companyName, websiteName, and url.

4. Create getters and setters for all of the JobPosting class's fields.

5. In the model package, create an LinkedinStrategy class that implements Strategy.
This class will implement a specific strategy for working with the Linkedin website (https://www.linkedin.com/jobs/).